requirements.txt
appdirs==1.4.3
bcrypt==3.1.7
cffi==1.14.0
cryptography==2.9
dataclasses==0.6
distlib==0.3.0
Django==2.2.6
django-crispy-forms==1.9.0
django-readonly-field==1.0.5
djangorestframework==3.11.0
djongo==1.2.38
filelock==3.0.12
paramiko==2.7.1
#pip==20.0.2
psycopg2==2.8.4
pycparser==2.20
pymongo==3.10.1
PyNaCl==1.3.0
pytz==2019.3
setuptools==41.2.0
six==1.14.0
sqlparse==0.2.4
sshtunnel==0.1.5
dacite==0.0.26
virtualenv==20.0.16
